+++ app/src/main/java/com/r0adkll/deckbuilder/arch/data/features/preview/RemotePreviewRepository.kt
@@ -1,8 +1,8 @@
 package com.r0adkll.deckbuilder.arch.data.features.preview
 
 import com.r0adkll.deckbuilder.arch.data.AppPreferences
-import com.r0adkll.deckbuilder.arch.data.remote.Remote
-import com.r0adkll.deckbuilder.arch.data.remote.model.ExpansionPreview
+import com.r0adkll.deckbuilder.arch.domain.features.remote.Remote
+import com.r0adkll.deckbuilder.arch.domain.features.remote.model.ExpansionPreview
 import com.r0adkll.deckbuilder.arch.domain.features.preview.PreviewRepository
 import com.r0adkll.deckbuilder.util.extensions.iso8601
 import io.reactivex.Observable