+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.7.8
+_IN PROGRESS_
+* Fixed issue with validation icon flashing
+* Fixed UI bug in expansion browser header
+
+
 # 1.7.7
 * Hotfix for issue where decks were missing cards due to find(ids) paging limit